Gateway extract
