+++ pipelex/cogt/extract/extract_worker_factory.py
@@ -19,13 +19,13 @@ class ExtractWorkerFactory:
         plugin_sdk_registry = get_plugin_manager().plugin_sdk_registry
         extract_worker: ExtractWorkerAbstract
         match plugin.sdk:
-            case "portkey_extract":
-                from pipelex.plugins.portkey.portkey_completions_factory import PortkeyCompletionsFactory  # noqa: PLC0415
-                from pipelex.plugins.portkey.portkey_extract_worker import PortkeyExtractWorker  # noqa: PLC0415
+            case "gateway_extract":
+                from pipelex.plugins.gateway.gateway_extract_worker import PortkeyExtractWorker  # noqa: PLC0415
+                from pipelex.plugins.gateway.gateway_factory import GatewayFactory  # noqa: PLC0415
 
                 extract_sdk_instance = plugin_sdk_registry.get_sdk_instance(plugin=plugin) or plugin_sdk_registry.set_sdk_instance(
                     plugin=plugin,
-                    sdk_instance=PortkeyCompletionsFactory.make_portkey_client(backend=backend),
+                    sdk_instance=GatewayFactory.make_portkey_client(backend=backend),
                 )
 
                 extract_worker = PortkeyExtractWorker(
